refactor(metadata): demote nixlMDConfig to its own header
Every field in the struct is backend-specific and only travels through the
manager because it arrives as a public agent setting, so sitting in the contract
header made it look like part of the contract: nixlMetadataBackend never takes
it, and only the three backends do. Move it to nixl_md_config.h and say there
that it is temporary - the direction is a backend reading its own parameters
from nixl::config, and the struct goes away once the fields can leave
nixlAgentConfig.
